refactor: nav 컴포넌트를 일반 컴포넌트로 분리한다
